refactor(PackageManagerTest): Move `findManagedFiles()` tests to `cli`
These tests create a tight coupling to pretty much all package manager
implementations. However, as package managers are implemented as plugins
and will be moved to separate projects which the analyzer must not depend
on, pragmatically move these tests to the CLI module which anyway bundles
all dependencies, and will in the future also depend on the package
manager plugins platform project.

While the `findManagedFiles()` tests are actually something like
integration tests, move them to a functional test as that source set
already exists and semantically comes closest.
